+++ package.json
@@ -26,7 +26,6 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.52.2",
-    "react-router-dom": "^6.26.1",
     "tailwind-merge": "^2.4.0",
     "zod": "^3.23.8"
   },
