+++ src/app/form/[id]/layout.tsx
@@ -4,7 +4,6 @@ export default async function FormLayout({
   children,
 }: Readonly<{
   children: React.ReactNode
-  params: { id: string }
 }>) {
   return (
     <div className='mx-auto flex w-full max-w-3xl flex-col items-center justify-center gap-8 py-8'>
